+++ webrtc/test/configurable_frame_size_encoder.cc
@@ -69,7 +69,7 @@ int32_t ConfigurableFrameSizeEncoder::Release() {
 }
 
 int32_t ConfigurableFrameSizeEncoder::SetChannelParameters(uint32_t packet_loss,
-                                                           int64_t rtt) {
+                                                           int rtt) {
   return WEBRTC_VIDEO_CODEC_OK;
 }
